fix login path error, employee global object and annoying alert in employee page
Once the user was login in, he was redirected to the wrong path because we were missing a `/` before the url. By adding the missing character, the user is now redirected to the dashboard as we wanted.
The global user object was reloaded in the dashboard componenet leading to an empty user. Since the global user object is loaded on both login and signin, removing the implementation in the Dashboard component fixes the empty user object.
Some errors, mostly linked to multiple useEffect firing, were insert inside an alert which was firing whenever something bad was happening. Removing the alert fixes the problem and improve the user experience.